Separate gauges into their own individual files
